backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
PyMySQL==1.1.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-dotenv==1.1.1
pydantic[email]==2.11.7